% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{attach_example200}
\alias{attach_example200}
\title{Attach example genotypes in the FBM format (bigstatsr).}
\usage{
attach_example200()
}
\description{
The function attaches simulated genotype matrix
with 1,500 samples (rows) and 200 genotypes (colums).
Genotypes are coded as 0, 1 and 2 (the number of alternatives alleles)
and all have the minor allele frequency 0.5.
}
\examples{
G <- attach_example200()
G
G[1:5, 1:10]

}
