% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglmmg.R
\name{big_scale_grm}
\alias{big_scale_grm}
\title{Scale and normalize genotypes for GRM.}
\usage{
big_scale_grm(center = TRUE, scale = TRUE, M)
}
\arguments{
\item{center}{Center?}

\item{scale}{Scale?}

\item{M}{A normalization value. 
It is typically equal to the number of columns in a matrix
(for which the scaling function is created).}
}
\value{
A function that returns a data.frame with two columns 
  "center" and "scale".
}
\description{
The function creates a scaling function 
to be used for linear algebra operations on FBMs
through the fun.scaling argument in downstream functions from bigstatsr.
}
\details{
The scaling function works such that the input matrix of raw genotypes
is represented as a matrix Z, 
which cross product is the genetic relationship matrix (GRM),
defined as scale(G) scale(G)' / M.
}
\examples{
G <- attach_example200()
dim(G)

M <- dim(G)[2] # number of columns
fun_sc <- big_scale_grm(M = M)
stats <- do.call(fun_sc, list(X = G))
str(stats)

}
