% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglmm.R
\name{biglmm}
\alias{biglmm}
\title{Low-rank LMM with a single random effect and residual random effect.}
\usage{
biglmm(y, X, Z, scale = FALSE, REML = TRUE, store_mat = FALSE,
  verbose = 0)
}
\description{
Model definition:
y_i =  X_i beta + u_i + e_i
}
\details{
u ~ N(0, s2*h2*G)
e ~ N(0, s2 I)

var(y) = V = s2 * (h2*G + I)
}
