% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglmmg.R
\name{biglmmg}
\alias{biglmmg}
\title{Low-rank LMM with a single random effect and residual random effect.}
\usage{
biglmmg(
  y,
  X,
  G,
  cols = seq(ncol(G)),
  M = length(cols),
  K = NULL,
  REML = TRUE,
  compute_mult = TRUE,
  verbose = 0
)
}
\description{
A version of biglmmz function without having to explicitly scale Z.
}
\details{
Model definition:
y_i =  X_i beta + u_i + e_i

u ~ N(0, s2*h2*G)
e ~ N(0, s2 I)

var(y) = V = s2 * (h2*G + I)
}
