% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowrank.R
\name{biglr_cprodMatInv}
\alias{biglr_cprodMatInv}
\title{Computes X'Vi}
\usage{
biglr_cprodMatInv(comp, Z, Xmat, K = NULL, transpose = FALSE)
}
\arguments{
\item{comp}{two-element vector of variance components}

\item{Z}{FBM of scaled genotypes}

\item{Xmat}{matrix of covariates}

\item{transpose}{(default FALSE) logical
whether return transposed product (X'Vi)' = Vi' X}
}
\value{
a product X'Vi
}
\description{
Computes X'Vi in a optimized way given that
 - Vi is the inverse of matrix V
 - V = comp[1]*G + comp[2]*diag(N)
 - G = tcrossprod(Z)
 - G is a low-rank matrix: ncol(Z) << nrow(Z)
}
