% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowrank.R
\name{biglr_det}
\alias{biglr_det}
\title{- V = comp[1]*G + comp[2]*diag(N)
- G = tcrossprod(Z)
- G is a low-rank matrix: ncol(Z) << nrow(Z)}
\usage{
biglr_det(comp, Z, K = NULL, log = TRUE)
}
\description{
(1) \url{https://en.wikipedia.org/wiki/Matrix_determinant_lemma}
|A + UCV| = |L| |C| |A|
where L = C- + VA-U
(2) For our specific case:
|D + ZHZ'| = |L| |H| |D|
D = diag(comp1), H = diag(comp2), L = (H- + Z'D-Z)
}
