% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{Compute the effective sample size for a LMM.}
\usage{
ess(G, cols = seq(ncol(G)), M = length(cols), h2, s2 = 1)
}
\arguments{
\item{G}{A FBM matrix of genotypes. Missing values are not handled.}

\item{cols}{A vector of columns of G to be used in the model.
By default, all columns of G are used.}

\item{M}{A scalar for normalization of the
 genetic relationship matrix: GRM = Z'Z / M,
 where Z is a matrix of standardized genotypes.
By default, M = length(cols).}

\item{h2}{The estimated heritability by LMM.}

\item{s2}{The estimated scaling constant for the variance components in LMM.}
}
\value{
A data.frame of results.
}
\description{
Compute the effective sample size for a LMM.
}
\examples{
G <- attach_example200()
G
G[1:5, 1:10]

ess(G, h2 = 0.5)

ess(G, h2 = 0.8)

}
