% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowrank.R
\name{lr_cprodMatInv}
\alias{lr_cprodMatInv}
\title{Computes X'Vi in an efficient way in a low-rank scenario (see below)}
\usage{
lr_cprodMatInv(comp, Z, X)
}
\description{
- Vi is the inverse of matrix V
 - V = comp[1]*G + comp[2]*diag(N)
 - G = tcrossprod(Z)
A low-rank scenario: G is a low-rank matrix,
i.e. ncol(Z) << nrow(Z)
}
