% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowrank.R
\name{naive_cprodMatInv}
\alias{naive_cprodMatInv}
\title{Computes X'Vi in a naive way}
\usage{
naive_cprodMatInv(comp, Z, X)
}
\description{
- Vi is the inverse of matrix V
 - V = comp[1]*G + comp[2]*diag(N)
 - G = tcrossprod(Z)
The focal scenario: G is a low-rank matrix,
i.e. ncol(Z) << nrow(Z)
}
