% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{attach_example200}
\alias{attach_example200}
\title{Attach example genotypes in the FBM format (bigstatsr).}
\usage{
attach_example200()
}
\description{
The matrix of 200 genotypes coded as 0, 1 and 2 (columns)
in 1,000 individuals (rows).
}
\examples{
G <- attach_example200()
G
G[1:5, 1:10]

}
