% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{attach_example200}
\alias{attach_example200}
\title{Load simulated genotypes in the FBM format.}
\usage{
attach_example200()
}
\value{
An object of class FBM.
}
\description{
The function loads the matrix of simulated genotypes
with 1,500 samples (rows) and 200 genetic variants (columns).
Genotypes are coded as 0, 1 and 2 (the number of alternative alleles).
All variants have the minor allele frequency fixed to 0.5.
}
\examples{
G <- attach_example200()
G
G[1:5, 1:10]

}
