% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglmmg.R
\name{big_scale_grm}
\alias{big_scale_grm}
\title{Scale and normalize the genotype matrix}
\usage{
big_scale_grm(center = TRUE, scale = TRUE, M)
}
\arguments{
\item{center}{Center?}

\item{scale}{Scale?}

\item{M}{A normalization value. It is typically equal to sqrt(M).}
}
\value{
A function that returns a data.frame with two columns 
  "center" and "scale".
}
\description{
The function creates a scaling function 
to be used for linear algebra operations on FBMs
through the fun.scaling argument in functions from bigstatsr.
}
\details{
The output of scaling function is such a matrix Z,
which cross product is the genetic relationship matrix (GRM),
defined as scale(G) scale(G)' / M
}
