% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biglmmz.R
\name{biglmmz}
\alias{biglmmz}
\title{Low-rank LMM with a single random effect and residual random effect.}
\usage{
biglmmz(
  y,
  X,
  Z,
  M,
  cols,
  backingfile = tempfile(),
  copy_Z = TRUE,
  K = NULL,
  scale = FALSE,
  impute = FALSE,
  REML = TRUE,
  compute_K = TRUE,
  store_mat = FALSE,
  verbose = 0
)
}
\description{
Model definition:
y_i =  X_i beta + u_i + e_i
}
\details{
u ~ N(0, s2*h2*G)
e ~ N(0, s2 I)

var(y) = V = s2 * (h2*G + I)
}
